+++ BooksStore/ViewModel/Responses/References/PainterStyles/ListPainterStylesResponse.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using ViewModel.Models.References;
 
-namespace ViewModel.Responses.References
+namespace ViewModel.Responses.References.PainterStyles
 {
     /// <summary>
     /// Ответ художественный стили художника
